# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "chat-API",
        "contact": "marco-2882@hotmail.com",
        "description": "This is a whatsApp clone api",
        "version": "1.0"
    },
    "paths": {
...
        "api/v1/users/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "login user",
                "operationId": "loginuser",
                "parameters": [],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/loginuserrequest"
                                    },
                                    {
                                        "example": {
                                            "email": "ian.rosas@gmail.com",
                                            "password": "xx1234"
                                        }
                                    }
                                ]
                            },
                            "example": {
                                "email": "ian.rosas@gmail.com",
                                "password": "xx1234"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false,
                "servers": [
                    {
                        "url": "http://localhost:8001",
                        "variables": {}
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "loginuserrequest": {
                "title": "loginuserrequest",
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "example": {
                    "email": "ian.rosas@gmail.com",
                    "password": "xx1234"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}